added test cases

+++ src/main/java/com/example/hospital/services/users/PatientService.java
@@ -1,14 +1,15 @@
 package com.example.hospital.services.users;
 
 import com.example.hospital.ResponseMessages;
+import com.example.hospital.exceptions.BadRequestException;
 import com.example.hospital.models.*;
+import com.example.hospital.models.enums.AppointmentStatus;
 import com.example.hospital.repositories.*;
 
 import java.util.List;
 
 import org.springframework.stereotype.Service;
 import jakarta.transaction.Transactional;
-
 
 @Service
 public class PatientService {
@@ -32,10 +33,10 @@ public class PatientService {
     @Transactional
     public Appointment bookAppointment(Appointment appointment) {
         Doctor doctor = doctorRepository.findById(appointment.getDoctor().getId()).orElseThrow(
-            () -> new IllegalArgumentException(ResponseMessages.record_not_found("Doctor"))
+            () -> new BadRequestException(ResponseMessages.record_not_found("Doctor"))
         );
         Patient patient = patientRepository.findById(appointment.getPatient().getId()).orElseThrow(
-            () -> new IllegalArgumentException(ResponseMessages.record_not_found("Patient"))
+            () -> new BadRequestException(ResponseMessages.record_not_found("Patient"))
         );
 
         appointment.setDoctor(doctor);
@@ -46,14 +47,19 @@ public class PatientService {
 
     public void cancelAppointment(Long appointmentId) {
         Appointment appointment = appointmentRepository.findById(appointmentId).orElseThrow(
-            () -> new IllegalArgumentException(ResponseMessages.record_not_found("Appointment"))
+            () -> new BadRequestException(ResponseMessages.record_not_found("Appointment"))
         );
+
+        if (appointment.getStatus() == AppointmentStatus.COMPLETED) {
+            throw new BadRequestException(ResponseMessages.CANNOT_CANCEL_COMPLETED_APPOINTMENT);
+        }
+
         appointmentRepository.delete(appointment);
     }
 
     public Appointment getAppointmentById(Long appointmentId) {
         return appointmentRepository.findById(appointmentId).orElseThrow(
-            () -> new IllegalArgumentException(ResponseMessages.record_not_found("Appointment"))
+            () -> new BadRequestException(ResponseMessages.record_not_found("Appointment"))
         );
     }
 
@@ -64,7 +70,7 @@ public class PatientService {
     public Review reviewDoctor(Review review) {
         boolean canReview = canPatientReviewDoctor(review.getPatient().getId(), review.getDoctor().getId());
         if (!canReview) {
-            throw new IllegalArgumentException(ResponseMessages.CANNOT_REVIEW_DOCTOR);
+            throw new BadRequestException(ResponseMessages.CANNOT_REVIEW_DOCTOR);
         }
 
         return reviewRepository.save(review);
